Changes requested in PR review by @ltalirz
- TIMEOUT_TIME --> TIMEOUT_SECONDS
- Updated exception message for `WrongCategory` in `validate_categories`
- Update design of index page.

The latter specifically concerns the titles "Description:" and "Categories:".
These are now removed.
Categories are shifted up inline with title.
If no categories are specified, an addition will be made to the description.
The same text will appear instead of categories on the single-entry site.

Additionally, if the `categories` key is specified for an app in `apps.json`,
and an empty list is the value, the same warning as if the key was not specified at all will be printed.
Only if a non-list or wrong category is specified, will a raise happen and the build fail.