aiidalab-home repo rename (#102)
After changing the branch name from `master` to `main` of aiidalab-home, the build failed with no branch find for the aiidalab-home. Fix by changing the registry of aiidalab-home the branch to main.

Pinning the `packaging<22`, to fix the error indicate the `stable` is not a valid version. It was caused by the new version of package packaging, I pinned it to <22 and it works. I can also reproduce the issue after updating the packaging to 22.0.