Adding external_url to the metadata for apps which are lacking this field (#113)
For some apps such as aiidalab-widgets-base the metadata is defined in the registry apps.yaml rather than read from the repository. It causes the issue that if the URL is not in the metadata, the issue aiidalab/aiidalab#329 is raised.

Instead of trying to push all apps using the new way to define metadata, at the moment, it is more proper to add the external_url to the metadata at the moment instead of moving to read metadata all from the app repository. We are not guaranteed that the metadata format of every app is correct, which will bring more work to us. So this change is more reliable.